+++ src/ftxui/dom/separator.cpp
@@ -18,16 +18,16 @@
 namespace ftxui {
 
 namespace {
-using Charset = std::array<std::string, 2>;  // NOLINT
-using Charsets = std::array<Charset, 6>;     // NOLINT
-// NOLINTNEXTLINE
-const Charsets charsets = {
-    Charset{"│", "─"},  // LIGHT
-    Charset{"╏", "╍"},  // DASHED
-    Charset{"┃", "━"},  // HEAVY
-    Charset{"║", "═"},  // DOUBLE
-    Charset{"│", "─"},  // ROUNDED
-    Charset{" ", " "},  // EMPTY
+using SeparatorCharset = std::array<std::string, 2>;  // NOLINT
+using SeparatorCharsets = std::array<SeparatorCharset, 6>;     // NOLINT
+
+const SeparatorCharsets charsets = { // NOLINT
+    SeparatorCharset{"│", "─"},  // LIGHT
+    SeparatorCharset{"╏", "╍"},  // DASHED
+    SeparatorCharset{"┃", "━"},  // HEAVY
+    SeparatorCharset{"║", "═"},  // DOUBLE
+    SeparatorCharset{"│", "─"},  // ROUNDED
+    SeparatorCharset{" ", " "},  // EMPTY
 };
 
 class Separator : public Node {